requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
plotly>=5.18.0
requests>=2.31.0
python-dotenv>=1.0.0
